Combine short flags if they're single runes
Any short flag that is a single rune is added to a list of okay flags to
combine, so you can do `-ab` as shorthand for `-a` `-b`.  Short flags
that are longer, or any argument that takes another arg (ie int, string)
are currently not allowed to be combined.

I based this off of the `fixInvoked` branch otherwise the tests wouldn't
pass.